shorts_factory/requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
google-auth>=2.22.0

# AI APIs  
google-generativeai>=0.7.0

# Media Processing
ffmpeg-python>=0.2.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.0

# Logging and Monitoring
colorlog>=6.7.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
